Fix targeting a bit
In the GPAM code the case of targeting a bit was not handled correctly
(it should not be class classification). Moreover it seems that
TensorFlow is now doing a type check for multiplication (needed by
metrics such as MeanRank).

+++ scaaml/io/dataset.py
@@ -553,7 +553,7 @@ class Dataset:
                 max_val = data["max_val"]
                 if max_val == 2:
                     # Binary classification.
-                    v = rec[data["ap"]][data["byte"]]
+                    v = tf.cast(rec[data["ap"]][data["byte"]], dtype=tf.float32)
                 else:
                     # Multiple classes classification.
                     v = tf.one_hot(rec[data["ap"]][data["byte"]], max_val)
